+++ Changelog.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Better command line argument validation
+- Better error messages
+- Support for multiple config files
 
 ## [1.2.0] - 2022-11-03